PSR-4ify the feature plugin and install classes.
