+++ src/API/Themes.php
@@ -81,8 +81,6 @@ class Themes extends \WC_REST_Data_Controller {
 
 		include_once ABSPATH . 'wp-admin/includes/file.php';
 		include_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
-		include_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-theme-upgrader.php';
-		include_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-theme-upgrader-skin.php';
 
 		$_GET['package'] = true;
 		$file_upload     = new \File_Upload_Upgrader( 'pluginzip', 'package' );
