+++ src/Notes/WC_Admin_Notes_Historical_Data.php
@@ -11,6 +11,8 @@ namespace Automattic\WooCommerce\Admin\Notes;
 
 defined( 'ABSPATH' ) || exit;
 
+use \Automattic\WooCommerce\Admin\WC_Admin_Install;
+
 /**
  * WC_Admin_Notes_Historical_Data.
  */
@@ -21,7 +23,7 @@ class WC_Admin_Notes_Historical_Data {
 	 * Creates a note for regenerating historical data.
 	 */
 	public static function add_note() {
-		$is_upgrading = get_option( \WC_Admin_Install::VERSION_OPTION );
+		$is_upgrading = get_option( WC_Admin_Install::VERSION_OPTION );
 		if ( $is_upgrading ) {
 			return;
 		}
